markdownMagicv2/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
markitdown==0.1.0
groq==0.4.2
python-dotenv==1.0.1